refactor. wish query hook useProduct 병합
