+++ src/query/useProduct.ts
@@ -1,9 +1,25 @@
 import type { AxiosError, AxiosResponse } from 'axios';
 import { useMemo } from 'react';
-import { useMutation, useQuery, useQueryClient } from 'react-query';
+import {
+  useInfiniteQuery,
+  useMutation,
+  useQuery,
+  useQueryClient,
+} from 'react-query';
+import { useLocation } from 'react-router-dom';
 import queryKeys from '../constants/queryKeys';
-import { getAdList, getList, postLike } from '../shared/api/productApi';
-import type { IAdResponse, IListResponse, TError } from '../shared/types/types';
+import {
+  getAdList,
+  getList,
+  getWish,
+  postLike,
+} from '../shared/api/productApi';
+import type {
+  IAdResponse,
+  IListResponse,
+  TError,
+  TResWish,
+} from '../shared/types/types';
 
 type TParameter = {
   page?: number;
@@ -20,6 +36,9 @@ const useProduct = ({
   keyword,
   productNo,
 }: TParameter) => {
+  const pathPattern = useLocation();
+  const [, path] = pathPattern.pathname.split('/');
+
   const { data: listData } = useQuery<AxiosResponse<IListResponse>, Error>(
     [queryKeys.LIST, category, sort, page],
     () => getList(category, sort, page),
@@ -70,6 +89,37 @@ const useProduct = ({
     },
   });
 
+  const wishQuery = useInfiniteQuery(
+    queryKeys.WISH,
+    ({ pageParam = 1 }) => getWish(pageParam),
+    {
+      getNextPageParam: (lastPage, allPages) => {
+        const maxPage = Math.ceil(lastPage.data.cnt / 12);
+        const nextPage = allPages.length + 1;
+        return nextPage <= maxPage ? nextPage : undefined;
+      },
+      enabled: path === queryKeys.WISH,
+      staleTime: 10 * 1000,
+      refetchOnWindowFocus: false,
+      select: data => ({
+        pages:
+          [...data.pages]
+            .map((list: AxiosResponse<TResWish>) => list?.data?.wishList)
+            .flat() || null,
+        pageParams: [...data.pageParams],
+      }),
+    }
+  );
+
+  const { wishCnt, wishData, wishLike } = useMemo(
+    () => ({
+      wishCnt: wishQuery?.data?.pages.length || 0,
+      wishData: wishQuery?.data?.pages || [],
+      wishLike: wishQuery?.data?.pages.map(list => list.p_No) || [],
+    }),
+    [wishQuery]
+  );
+
   return {
     totalPages,
     products,
@@ -79,6 +129,10 @@ const useProduct = ({
     adProducts,
     likeProducts,
     mutate,
+    wishQuery,
+    wishCnt,
+    wishData,
+    wishLike,
   };
 };
 export default useProduct;
